py-srv/bin/requirements.txt
bottle
bottle-sqlalchemy
sqlalchemy-cockroachdb
psycopg2